pyproject.toml
[project]
name = "api-dynamic-ingestion"
version = "0.1.0"
description = "API Dynamic Ingestion"
authors = [
    {name = "jparep",email = "parepson@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "snowflake-connector-python (>=3.14.0,<4.0.0)",
    "dotenv-python (>=0.0.1,<0.0.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
